Add logic to update the bucket policy for distributing the image copier
This will add a bucket policy allowing all of our other accounts to access the part of the bucket to which we upload the lambda code allowing us to more easily update the lambda across the estate and also supports uploading to the right place at deploy time.

Actually updating the lambdas is not yet done.