Switch easy_install to pip
`easy_install` is no longer available on Ubuntu 18.04 but it's fine to
use `pip` to install the package. Unfortunately the package depends on
a library that does not work under Python 3 so we're forced to use
`pip` rather than `pip3`. This means on Debian we need to ensure that
pip is installed.