Remove Mutex > Webserver dependency
Originally it was thought the webserver role could be set after
mutex. Unfortunately, the webserver role is needed to yield the ELB
health check on :8080/health as it provides the resources to allow
requests in that port (gunicorn etc). Don't make the mistake of having
mutex before webserver or builds will not report a positive health
check for the node while there are two in the ASG (condition of mutex)
so RiffRaff will not know the old one is ready for kill.