Prepend AIRFLOW_HOME to PYTHONPATH
There are many duplications and antipatterns in the Airflow setup. But
this PYTHONPATH setting meanst that AIRFLOW_HOME isn't known to the
PYTHON executable when it is invoked by the systemd services that manage
the webserver, executor and scheduler. That is because in the BAKE this
file is  unpacked into /etc/sysconfig/ which is then used for the
Environment File.

Just adding a second value means that we can use conventions in Airflow
development that import from the path. Such as imports from DAG files
without relative imports.