Add a link to view the AMI in AMIable.
Deep-link into AMIable. This view in AMIable provides information about the AMI, including:
- creation date
- ARN
- any launch configurations that use it
- any instances that use it
- any available updates

This is useful to double check if we have setup a stack's auto-update correctly: if we see the most recent AMI not used, but an AMI from last month is, then we've likely misconfigured the app in RiffRaff.