Make secret fetch fail on error
Ansible uses the return code of a script to decide wether a shell task failed.

For some reason, when `aws secretsmanager get-secret-value` fails it
keeps its return code set to zero, so the task isn't marked as a
failure and the bake continues.

Luckily, ansible has a built in function to fail based on an error
message, and in the examples I've seen there's always the word
"error" in the error message.

```
* An error occurred (AccessDeniedException) when calling the GetSecretValue operation:
* aws: error: argument --region: expected one argument
```