refactor: Do not download from S3 unnecessarily
We're currently downloading the same content from S3 (bucket teamcity-ivy-cache) into two locations:
  1. /opt/teamcity/.ivy2/cache/
  2. /opt/teamcity/buildAgent/system/sbt_ivy/cache/

This change removes the second download, replacing it with a local disk copy command.
This results in less network IO and a faster AMI bake.
Logs from previous bakes suggest it takes ~3 minutes to perform the S3 download.

It should also mean a reduction in S3 spend, however looking at the billing dashboard,
the savings will be marginal as current spend is in the tens of dollars, so not massive amounts.