feat: Define `SSMRunCommandPolicy` in the CDK stack
Moving the `SSMRunCommandPolicy` resource into the CDK stack.

This is chosen because it's fairly easy to test if it works: if successful, we can continue to connect to an instance using ssm-scala.

AMIgo's stack defines two additional actions over the standard `GuSSMRunCommandPolicy`:
  1. ssm:StartSession
  2. ssm:TerminateSession

Slightly annoyingly, `GuSSMRunCommandPolicy` is a singleton and doesn't accept additional input.
Therefore, we have to add these actions into a new statement on the policy.

`GuSSMRunCommandPolicy`'s `Roles` array is kinda hilarious (see snapshot test change).
The good news is it's done by AWS CDK magic, so should work!