feat: Define EC2 policies with GuCDK construct
An app's identity is defined by the App, Stack and Stage tags.

This change uses the `GuDescribeEC2Policy` construct to allow an app to know it's identity,
replacing the YAML definition.

In using the construct, we're also creating the policy as a resource in its own right.
This is preferred as it follows the single responsibility principle.

The `GuDescribeEC2Policy` construct grants the following permissions on `"*"`:
  - autoscaling:DescribeAutoScalingInstances
  - autoscaling:DescribeAutoScalingGroups
  - ec2:DescribeTags
  - ec2:DescribeInstances

That is, there is an additional `ec2:DescribeInstances` permission being granted.