feat: Move Anghammarad policy into CDK
This change moves the Anghammarad permissions into the CDK defined stack.

The permissions get created as a policy in its own right,
rather than statements on a generic "amigo-app" policy.

This results in a LOC increase to the resulting CFN template.
That is, we're increasing the chance of reaching the CFN template limit!

This stack is pretty small, we should adjust strategy only when needed.