feat: Use the `GuPlayApp` pattern
This change uses the `GuPlayApp` pattern to define the ASG, ALB etc for AMIgo.

With this change, we're creating a new load balancer to run in parallel to the current one.
We can then perform a DNS migration and start removing the YAML template to have a CDK only stack!