Downgrade puppeteer version back to 1.9.0
When time machine stopped working, we raised this PR to address the
problem:
https://github.com/guardian/amigo/pull/644

However, once we deployed the changes to time machine, time machine was no
longer working. We wanted to see if we could upgrade puppeteer while we were
upgrading libgtk anyway and included it in the PR. We tested upgrading both
versions on the time machine box in PROD, process described here:

https://github.com/guardian/ophan/issues/3932#issuecomment-904571290

However, that doesn't seem to correspond with the changes we introduced in
Amigo and we'd like to decouple the Puppeteer upgrade from the libgtk upgrade.

Co-authored-by: Simone Smith  <simone.smith@guardian.co.uk>