chore: Remove `aws-lambda-java-log4j` dependency
This is part of the work to resolve the log4j remote code execution vulnerability.

`aws-lambda-java-log4j` can be used to log to ELK, however this is handled by
https://github.com/guardian/cloudwatch-logs-management already.

`cloudwatch-logs-management` will forward CloudWatch Logs to Central ELK.
Therefore, the only requirement for a lambda to log into Central ELK is for it
to write logs to CloudWatch Logs.

This is the default behaviour - `println`ing a message will be shipped to ELK
automagically.