Update metricbeat role to support ARM architecture
This switches from directly specifying the download url for the
debian package, to using Elastic's APT repository, in the hope that
APT will correctly download the correct architecture - ie ARM/AArch64
when we want it!

https://www.elastic.co/guide/en/beats/metricbeat/7.17/setup-repositories.html
https://github.com/guardian/amigo/pull/377#issuecomment-1064983514

The configuration in this commit- adding an APT repository - is
based on the configuration in the `fluentbit` role.