Replace Vagrant with Multipass
Vagrant is great but Virtualbox, which it uses under the hood, does not
support arm64 (including M1 macs) and will not in the near future.

Given the department is migrating to M1 macs shortly (and some
already have), we need another tool. Multipass
(https://multipass.run) is a good candidate here.