Migrate Image Copier stacks to CDK
Note, we have not used the GuSnsLambda pattern for now as it does
not permit a shared (across accounts) bucket for the lamdba file.

Also note, some of the parameters have been kept for now. This is
non-idiomatic in CDK but makes sense here as the stacks are stack-
setted and account-level overrides are possible.

Lastly, in the medium term we hope to delete these stacks anyway
and share encrypted AMIs directly.