Attempt (crude) fix of bake deletion housekeeping job
My guess is that the behaviour of 'limit' changed when Scanamo
was updated at the end of March 2022 - a timeframe that corresponds
to when we start seeing this fail.

It looks like Scanamo 'fixed' its behaviour to match what 'limit'
means in Dynamo parlance. For more info, see:

- https://github.com/scanamo/scanamo/pull/996
- https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Scan.html#Scan.Limit

Note, this is more to confirm the source of the issue for now. A
better long-term fix is to create a secondary index on the deleted
field and then query against that to avoid the scans here.