Add secondary index for 'deleted' to avoid scanning
Scanning for deleted items requires a table scan and has been the
source of a recent subtle bug:

https://github.com/guardian/amigo/pull/846

This commit adds a secondary index for deleted so that it becomes
easy to query for where deleted=true.

Note, we will need to add the index manually as the existing code
only creates the table (and any indexes) if it does not yet exist.