Remove limit from delete query
I'm not normally comfortable with this, but it's the only way to
ensure we retrieve all deleted=true items without changing the
data model (which would involve a good deal of work).

In this case, the number of table items is relatively small (100s
rather than 1000s+) and expected to remain so, so I don't think
this is a problem.

As a reminder, filter is applied to returned results, so if used
with a limit it is not possible to guarantee we reach the items
where deleted=true.

E.g.

filter: deleted=true
limit: 10

Imagine there are 11 items in the table and only one has deleted
set to true. Our scan is limited to 10 items by the limit, and only
*then* is the filter applied. So it's possible the deleted item is
never returned here!