Create `amigo.gutools.co.uk`
As part of the work to restore pre-incident URLs, this change creates `amigo.gutools.co.uk`.
It is a CNAME to the same ALB hosting `public.amigo.gutools.co.uk`.

The TLS certificate used by the ALB is also changed, adding the new domain as a subject alternative domain.

In order to offer seamless DX, we also configure a 301 redirect from the legacy domain to this new one.

We will remove the legacy domain in a future change.

The above is repeated for the CODE forms too.