Add wildcard to version on docker role
This is required as versions available beyond bionic include the os name in the version, in order for this role to remain agnostic and allow use on focal and beyond adds a wildcard to the version. This has been tested on bionic and build d7080c1 is installed with and without the wildcard, in other words this change should not impact existing recipes based on bionic.