`aws-tools` role uses latest awscli
Changes the `awscli` install step to always install the latest version. Essentially the inverse change to https://github.com/guardian/amigo/pull/549 where the `awscli` was pinned to a specific version due to a lack of support for Python 3.4/3.5

Attempts to solve a problem where the awscli was failing to install correctly. The problem was fixed in a recent version of the `awscli`: https://github.com/aws/aws-cli/issues/8036#issuecomment-1640932993