feat: consolidate dependency updates into weekly PR
We now have the following RFC for dependency management:

https://docs.google.com/document/d/1FIfovh2p08lWkXk3i83H4TAxf5V3Jj0j7wEUEe09Wcs/edit

It advocates individual PRs (raised by Scala Steward, Dependabot,
etc.) that are combined into a weekly PR.

There are a few ways we could do this, but here we:

* configure Dependabot and SS to raise PRs asap for individual
  updates
* group them into a weekly PR using github/combine-prs

Grouping is done for any PRs with the 'dependencies' label. Only
PRs that pass CI are combined.

Note, the daily individual updates will create noise in our PR
announcer chat messages, so if we choose this solution we should
ignore PRs with the 'dependencies' label in the announcer.