Add option to not automatically start fluentbit
Currently if you want or need to add extra fluentbit config at instance start time
(ie. in user data script) you must `systemctl restart td-agent-bit.service`, which
means that the cloud-init logs up to that point will be resent from the start, meaning
some logs are sent twice! If instead we never add the start-fluentbit script, we can
start it when we have the full config in place, and still benefit from `devx-logs`!