The recipe `hasUsage` function now includes launch template checks
Launch template support was added last year, but it doesn't look like
the UI was fully updated to reflect this. This means a manual tidy up
job might accidentally remove recipes that are in use, even if there
are not currently any running instances.

We add this check to the logic, so that the "Unused recipes" section
of the recipes index is correctly populated. This also improves the
usage checks on base images page.