Set up a queue for processing bakes async and replace cron scheduling
Currently Amigo will rebake recipes based on a Quartz cron schedule.
Unfortunately, bakes can fail or timeout if too many are running at
once, which means that users sometimes need to hunt for a free slot, or
adjust the scheduling of many recipes to get them all to play nicely.

We've had similar problems in the Grid when people try to upload many
images all in one go. The architectural solution we've implemented there
is to queue jobs, allowing the app to restrict its concurrency and limit
the number of jobs running at any one time.

Once we're doing this though, there seems little need for the precision
of a Cron expression - we always schedule weekly bakes, so besides the
choice of day do we really care exactly what hour/minute it starts?
Instead, only prompt the user to select a day, and then shortly after
midnight, have Amigo find and schedule the bakes for that day's recipes.